requirements.txt
sanic==23.3.0
uvloop==0.17.0
tortoise-orm==0.17.8
aiomysql==0.2.0
arq==0.25.0
Sanic-Cors==0.10.0.post3
Sanic-Plugins-Framework==0.9.2